feat: path 이름 추가
