+++ src/main/kotlin/com/brtrip/path/controller/request/PathRequest.kt
@@ -1,6 +1,5 @@
 package com.brtrip.path.controller.request
 
-import com.brtrip.path.domain.Path
 import com.brtrip.place.PlaceRequest
 
 class PathRequest(
