+++ src/main/kotlin/com/brtrip/path/controller/response/PathResponse.kt
@@ -6,13 +6,15 @@ import com.brtrip.place.dto.PlaceResponse
 
 class PathResponse(
     val places: List<PlaceResponse>,
-    val likeCount: Long
+    val likeCount: Long,
+    val name: String
 ) {
     companion object {
         fun of(path: Path, places: List<Place>): PathResponse {
             return PathResponse(
                 places = places.map { PlaceResponse.of(it) },
-                likeCount = path.likeCount
+                likeCount = path.likeCount,
+                name = path.name
             )
         }
     }
