+++ src/main/kotlin/com/brtrip/path/domain/Path.kt
@@ -1,7 +1,6 @@
 package com.brtrip.path.domain
 
 import com.brtrip.common.BaseEntity
-import com.brtrip.path.domain.PathPlace
 import javax.persistence.*
 
 @Entity
@@ -11,7 +10,10 @@ class Path(
     var id: Long? = null,
 
     @Column(name = "like_count")
-    var likeCount: Long = 0
+    var likeCount: Long = 0,
+
+    @Column(name = "name")
+    var name: String = ""
 ): BaseEntity() {
     @OneToMany(mappedBy = "path", cascade = [CascadeType.ALL], fetch = FetchType.LAZY)
     var pathPlaces: MutableList<PathPlace> = mutableListOf()
